Util: Convert Random to a virtual class
You now ask for a Random, and optionally what kind, but the
implementation is hidden.

Seeding and basic value generation is now done as u32 instead of s32.

One other change that was required here, is to store Building
spriteOffsets as a u16. We crash horribly if they actually try to use a
negative number, and the save format uses u16, so storing them as a
signed type was never correct - it just happened to be because our
random number generator only returned positive numbers from the
function that was used.

+++ src/Sim/Building.h
@@ -175,7 +175,7 @@ struct Building {
     Optional<s16> variantIndex;
 
     Entity* entity;
-    s32 spriteOffset; // used as the offset for getSprite
+    u16 spriteOffset; // used as the offset for getSprite
 
     s32 currentResidents;
     s32 currentJobs;
